Use close_range when possible
