+++ crates/libcontainer/Cargo.toml
@@ -18,8 +18,8 @@ wasm-wasmer = ["wasmer", "wasmer-wasi"]
 
 [dependencies]
 anyhow = "1.0"
+bitflags = "1.3.2"
 caps = "0.5.3"
-
 chrono = { version="0.4", features = ["serde"] }
 crossbeam-channel = "0.5"
 dbus = "0.9.5"
@@ -37,6 +37,7 @@ libcgroups = { version = "0.0.2", path = "../libcgroups" }
 libseccomp = { version = "0.0.2", path = "../libseccomp" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+syscalls = "0.5.0"
 rust-criu = { git = "https://github.com/checkpoint-restore/rust-criu", version = "0.1.0" }
 wasmer = { version = "2.1.1", optional = true }
 wasmer-wasi = { version = "2.1.1", optional = true }
